refactor(plugins): split runtime injectors into domain plugins
