+++ assistant/src/__tests__/injector-registry-order-guard.test.ts
@@ -1,23 +1,32 @@
 /**
  * Guard test: the injector registry materializes the exact same ordered
- * sequence the legacy hard-coded chain produced.
+ * sequence the legacy hard-coded chain produced, now that the default injectors
+ * are split across the memory plugin and the domain plugins (`turn-context`,
+ * `workspace`, `documents`, `channel`, `session`).
  *
  * The per-turn injection chain is a registry union (see
- * `plugins/injector-registry.ts`) rather than a hard-coded import, so this test
- * locks the produced order: the full ordered id sequence is asserted both
- * against the recomputed sort of `defaultMemoryPlugin`'s contributed injectors
- * and against a hardcoded literal snapshot. Any reordering — a changed `order`,
- * a new injector, a regression in the registry's sort — fails here loudly
- * rather than silently shifting what the model sees per turn.
+ * `plugins/injector-registry.ts`) sorted by `order`, so this test locks the
+ * produced order: the full ordered id sequence is asserted both against the
+ * recomputed sort of every default plugin's contributed injectors and against a
+ * hardcoded literal snapshot. Any reordering — a changed `order`, a new
+ * injector, an injector moved between plugins, a regression in the registry's
+ * sort — fails here loudly rather than silently shifting what the model sees
+ * per turn.
  */
 
 import { afterEach, beforeEach, describe, expect, test } from "bun:test";
 
-import { defaultInjectors } from "../plugins/defaults/memory/injectors.js";
+import { channelInjectors } from "../plugins/defaults/channel/injectors.js";
+import { documentsInjectors } from "../plugins/defaults/documents/injectors.js";
+import { registerDefaultPluginInjectors } from "../plugins/defaults/index.js";
+import { memoryInjectors } from "../plugins/defaults/memory/injectors.js";
 import {
   memoryV3Injector,
   memoryV3SpotlightInjector,
 } from "../plugins/defaults/memory/v3/injector.js";
+import { sessionInjectors } from "../plugins/defaults/session/injectors.js";
+import { turnContextInjectors } from "../plugins/defaults/turn-context/injectors.js";
+import { workspaceInjectors } from "../plugins/defaults/workspace/injectors.js";
 import {
   clearInjectorRegistry,
   getRegisteredInjectors,
@@ -25,8 +34,17 @@ import {
 } from "../plugins/injector-registry.js";
 import type { Injector } from "../plugins/types.js";
 
-const CONTRIBUTED = [
-  ...defaultInjectors,
+// Every injector every default plugin contributes — the registry-independent
+// union, gathered straight from the plugins' exported arrays. Sorting this by
+// `order` must reproduce what `getRegisteredInjectors()` returns once the
+// defaults are registered.
+const ALL_CONTRIBUTED = [
+  ...turnContextInjectors,
+  ...workspaceInjectors,
+  ...documentsInjectors,
+  ...channelInjectors,
+  ...sessionInjectors,
+  ...memoryInjectors,
   memoryV3Injector,
   memoryV3SpotlightInjector,
 ];
@@ -56,7 +74,7 @@ const EXPECTED_ORDER = [
 describe("injector registry order guard", () => {
   beforeEach(() => {
     clearInjectorRegistry();
-    registerPluginInjectors("default-memory", CONTRIBUTED);
+    registerDefaultPluginInjectors();
   });
 
   afterEach(() => {
@@ -73,9 +91,9 @@ describe("injector registry order guard", () => {
 
   test("the materialized id sequence matches the recomputed sort and the snapshot", () => {
     const names = getRegisteredInjectors().map((i) => i.name);
-    // (a) matches the recomputed stable sort of the contributed injectors.
+    // (a) matches the recomputed stable sort of every contributed injector.
     expect(names).toEqual(
-      [...CONTRIBUTED].sort((a, b) => a.order - b.order).map((i) => i.name),
+      [...ALL_CONTRIBUTED].sort((a, b) => a.order - b.order).map((i) => i.name),
     );
     // (b) matches the hardcoded snapshot — the deliberate lock.
     expect(names).toEqual(EXPECTED_ORDER);
