+++ assistant/src/__tests__/memory-v2-static-injector.test.ts
@@ -37,14 +37,14 @@ mock.module("../config/loader.js", () => ({
   }),
 }));
 
-const { defaultInjectors } =
+const { memoryInjectors } =
   await import("../plugins/defaults/memory/injectors.js");
 import type { Injector, TurnContext } from "../plugins/types.js";
 import type { Message } from "../providers/types.js";
 import { getWorkspacePromptPath } from "../util/platform.js";
 
 function findInjector(name: string): Injector {
-  const injector = defaultInjectors.find((i) => i.name === name);
+  const injector = memoryInjectors.find((i) => i.name === name);
   if (!injector) {
     throw new Error(`injector '${name}' not registered`);
   }
