+++ assistant/src/__tests__/plugin-import-boundary-guard.test.ts
@@ -48,6 +48,7 @@ const DEFAULTS_ABS = join(process.cwd(), DEFAULTS_REL);
  * and paste the printed object here, explaining the new coupling in your PR.
  */
 const BASELINE: Record<string, readonly string[]> = {
+  channel: ["../../types.js", "../injector-order.js"],
   compaction: [
     "../../../config/loader.js",
     "../../../config/schemas/compaction.js",
@@ -63,6 +64,11 @@ const BASELINE: Record<string, readonly string[]> = {
     "../../../runtime/actor-trust-resolver.js",
     "../../../util/logger.js",
     "../../types.js",
+  ],
+  documents: [
+    "../../../documents/document-comments-store.js",
+    "../../types.js",
+    "../injector-order.js",
   ],
   "exploration-drift": ["../../../../subagent/index.js"],
   "image-fallback": ["node:crypto", "node:fs", "node:path"],
@@ -116,31 +122,32 @@ const BASELINE: Record<string, readonly string[]> = {
     "../../../../util/truncate.js",
     "../../../config/loader.js",
     "../../../context/strip-injections.js",
-    "../../../daemon/conversation-registry.js",
-    "../../../daemon/conversation-runtime-assembly.js",
-    "../../../daemon/conversation-workspace.js",
-    "../../../daemon/now-scratchpad.js",
     "../../../daemon/pkb-context-tracker.js",
     "../../../daemon/pkb-reminder-builder.js",
     "../../../daemon/trust-context.js",
-    "../../../documents/document-comments-store.js",
     "../../../memory/graph/conversation-graph-memory.js",
     "../../../memory/pkb/autoinject.js",
     "../../../memory/pkb/context.js",
     "../../../memory/pkb/pkb-search.js",
     "../../../memory/pkb/types.js",
     "../../../memory/v2/static-context.js",
-    "../../../runtime/capabilities.js",
     "../../../types.js",
     "../../../util/logger.js",
     "../../../util/platform.js",
     "../../types.js",
+    "../injection-presence.js",
+    "../injector-order.js",
     "@qdrant/js-client-rest",
     "node:crypto",
     "node:fs",
     "node:path",
     "uuid",
     "zod",
+  ],
+  session: [
+    "../../../config/loader.js",
+    "../../types.js",
+    "../injector-order.js",
   ],
   "task-progress-nudge": [
     "../../../../daemon/conversation-registry.js",
@@ -149,8 +156,27 @@ const BASELINE: Record<string, readonly string[]> = {
   ],
   "title-generate": [
     "../../../../config/loader.js",
-    "../../../../persistence/conversation-crud.js",
     "../../../../memory/conversation-title-service.js",
+    "../../../../persistence/conversation-crud.js",
+  ],
+  "turn-context": [
+    "../../../daemon/conversation-runtime-assembly.js",
+    "../../../runtime/capabilities.js",
+    "../../types.js",
+    "../injector-order.js",
+  ],
+  workspace: [
+    "../../../config/loader.js",
+    "../../../context/strip-injections.js",
+    "../../../daemon/conversation-registry.js",
+    "../../../daemon/conversation-workspace.js",
+    "../../../daemon/now-scratchpad.js",
+    "../../../daemon/trust-context.js",
+    "../../../util/platform.js",
+    "../../types.js",
+    "../injection-presence.js",
+    "../injector-order.js",
+    "node:fs",
   ],
 };
