+++ assistant/src/plugins/defaults/channel/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "default-channel",
+  "version": "1.0.0",
+  "description": "First-party default plugin contributing the Slack channel runtime injectors (chronological transcript and active-thread focus).",
+  "private": true,
+  "license": "MIT",
+  "type": "module",
+  "engines": {
+    "node": ">=20.12.0"
+  },
+  "peerDependencies": {
+    "@vellumai/plugin-api": "^0.8.0"
+  }
+}
