+++ assistant/src/plugins/defaults/session/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "default-session",
+  "version": "1.0.0",
+  "description": "First-party default plugin contributing the session-state runtime injectors (background-turn framing and subagent status).",
+  "private": true,
+  "license": "MIT",
+  "type": "module",
+  "engines": {
+    "node": ">=20.12.0"
+  },
+  "peerDependencies": {
+    "@vellumai/plugin-api": "^0.8.0"
+  }
+}
